+++ modules/portal-jslib/src/test/java/com/enonic/xp/portal/jslib/impl/GeneralUrlHandlerTest.java
@@ -3,6 +3,7 @@ package com.enonic.xp.portal.jslib.impl;
 import org.junit.Test;
 
 import com.enonic.wem.script.command.CommandHandler;
+import com.enonic.xp.portal.jslib.AbstractHandlerTest;
 
 public class GeneralUrlHandlerTest
     extends AbstractHandlerTest